refactor: move queue contract to neutral boundary
