+++ scripts/generate_inventory_docs.py
@@ -126,6 +126,7 @@ TOP_LEVEL_MODULE_PURPOSES = {
     "main.rs": "Binary entry point. Dispatches CLI commands or boots the server runtime.",
     "phase_gate.rs": "Immutable typed phase-gate declarations and snapshot compatibility checks shared by HTTP, policy dispatch, and durable reducers.",
     "pipeline.rs": "Pipeline stage loading, resolution, and transition helpers.",
+    "queue_contract.rs": "Queue field compatibility contracts shared by planner prompts, API documentation, and runtime-facing consumers.",
     "receipt.rs": "Receipt parsing and workspace attribution helpers.",
     "reconcile.rs": "Boot-time reconciliation for persisted state and dispatch-runtime drift.",
     "manual_intervention.rs": "Manual intervention parsing and helpers shared by Discord reply/requeue flows.",
